Always run composer install (even if a lock file is present)
The composer_project install action should always run, even if
there is a composer.lock file present in the project directory.
This will generally be the case on build and production machines
where composer is being used to deploy known versions of
dependencies.

composer install is idempotent (for the same lock file) so there
are no problems with running it multiple times on the same
project path.